Kunde kan nu klikke på knap og se stykliste efter de har betalt
