+++ src/main/resources/templates/payed.html
@@ -18,6 +18,12 @@
     <p>Du vil modtage en bekræftelsesmail med yderligere detaljer.</p>
     <p>Har du spørgsmål, er du velkommen til at kontakte os.</p>
     <br>
+
+    <form th:action="@{/payed/show-bom}" method="get">
+        <button class="custom-btn" type="submit">Se stykliste</button>
+    </form>
+
+
 </div>
 </body>
 </html>
